Applied PR updates
